% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_results}
\alias{clean_results}
\title{Clean unused sensitive data.}
\usage{
clean_results(dir_result)
}
\description{
A function ask user if they want to remove unused sensitive data once they have finished all calculation.
The function will find the unused sensitive data in current work directory.
}
