% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_statistics_tables.R
\name{compare_nordcan_statistics}
\alias{compare_nordcan_statistics}
\title{a wrapper for comparing current statistics with archived statistics}
\usage{
compare_nordcan_statistics(stats_current, stats_archived, ds_nms)
}
\description{
a wrapper for comparing current statistics with archived statistics
}
