% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_by_entity}
\alias{compute_by_entity}
\alias{compute_by_entity_column}
\title{Stratification by Entity}
\usage{
compute_by_entity(x, entities, fun, arg_list)

compute_by_entity_column(x, entities, fun, arg_list)
}
\arguments{
\item{x}{\verb{[data.table]} (mandatory, no default)

a NORDCAN cancer record dataset}

\item{entities}{\verb{[integer]} (mandatory, no default)

an integer vector of entity numbers}

\item{fun}{\verb{[function]} (mandatory, no default)

function that is applied to each entity / entity column; see Details}

\item{arg_list}{\verb{[list]} (mandatory, no default)

a \code{list} containing arguments to pass to \code{fun}; see Details}
}
\value{
Either a \code{list} or a \code{data.table}. Initially results are collected into a
\code{list} where each element is the set of results for an individual entity
or entity column. If every element of this list is a \code{data.table},
then every element is combined into one long \code{data.table} where column
\code{"entity"} will identify the entity. In the case of
\code{compute_by_entity} the column \code{"entity"} is created and in the case of
\code{compute_by_entity_column} each entity column is renamed to \code{"entity"}
in the entity column-specific results before combining.

\code{fun} is called using \code{arg_list} after the appropriate entity- or
entity column-specific modification using \verb{[base::do.call]}.

\code{compute_by_entity} modifies an element in \code{arg_list} named
\code{x}, intended to contain the dataset cancer record. \code{x} is subset into
each entity number separately before calling \code{fun}.

\code{compute_by_entity_column} modifies or creates \code{arg_list[["by"]]} before
calling \code{fun}. \code{x} is not subset by this function. Therefore
\code{compute_by_entity_column} is a more efficient but also less flexible
version of \code{compute_by_entity}.
}
\description{
Compute statistics by entity number.
}
\details{
\code{compute_by_entity} effectively loops over each entity number given in
\code{entities} and computes the results separately for each.

\code{compute_by_entity_column} loops over each column defining entities in
\code{x}. For each such column, \code{arg_list[["by"]]} is created or modified to have
(also) the column in question, limited to only those entities supplied in
\code{entities}.
}
