% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nordcanstat_metadata_statistics_tables_names}
\alias{nordcanstat_metadata_statistics_tables_names}
\alias{entity_strata}
\alias{loop_over_entity_numbers}
\alias{loop_over_entity_columns}
\alias{nordcanstat_by_entity}
\title{Stratification by Entity}
\usage{
nordcanstat_metadata_statistics_tables_names(
  participant_name = nordcancore::nordcan_metadata_participant_info()[["name"]]
)

loop_over_entity_numbers(x, entities, fun, arg_list)

loop_over_entity_columns(x, entities, fun, arg_list)

nordcanstat_by_entity(
  entities,
  arg_list,
  basicepistats_fun,
  loop_over = c("entity_numbers", "entity_columns")[1L]
)
}
\arguments{
\item{participant_name}{\verb{[character]} (optional, default taken from settings)

see \link[nordcancore:set_global_nordcan_settings]{nordcancore::set_global_nordcan_settings}}

\item{x}{\verb{[data.table]} (mandatory, no default)

a NORDCAN cancer record dataset}

\item{entities}{\verb{[NULL, integer]} (optional, default \code{NULL})

see \code{by} for how to stratify output by entity number.
\itemize{
\item \code{NULL}: output of
\code{nordcancore::nordcan_metadata_entity_no_set("all")}
is used
\item \code{integer}: these entity numbers are used
}}

\item{fun}{\verb{[function]} (mandatory, no default)

function that is applied to each entity / entity column; see Details}

\item{arg_list}{\verb{[list]} (mandatory, no default)

a \code{list} containing arguments to pass to \code{fun}; see Details}

\item{basicepistats_fun}{\verb{[function]} (mandatory, no default)

function from \strong{basicepistats}, e.g. \verb{[basicepistats::stat_count]};
this function will be run to produce results by entity number}

\item{loop_over}{\verb{[character]} (mandatory, default \code{"entity_numbers"})
\itemize{
\item \code{"entity_columns"}: function \code{loop_over_entity_columns} is called internally
\item \code{"entity_numbers"}: function \code{loop_over_entity_numbers} is called internally
}}
}
\value{
Either a \code{list} or a \code{data.table}. Initially results are collected into a
\code{list} where each element is the set of results for an individual entity
or entity column. If every element of this list is a \code{data.table},
then every element is combined into one long \code{data.table} where column
\code{"entity"} will identify the entity. In the case of
\code{loop_over_entity_numbers} the column \code{"entity"} is created and in the case of
\code{loop_over_entity_columns} each entity column is renamed to \code{"entity"}
in the entity column-specific results before combining.
}
\description{
Compute statistics by entity number.
}
\details{
\code{loop_over_entity_numbers} effectively loops over each entity number given in
\code{entities} and computes the results separately for each.

\code{loop_over_entity_columns} loops over each column defining entities in
\code{x}. For each such column, \code{arg_list[["by"]]} is created or modified to have
(also) the column in question, limited to only those entities supplied in
\code{entities}.

Both of these two functions are intended to be applied internally
by the \code{nordcanstat_} family of functions and not by the user.

\code{fun} is called using \code{arg_list} after the appropriate entity- or
entity column-specific modification using \verb{[base::do.call]}.

\code{loop_over_entity_numbers} modifies an element in \code{arg_list} named
\code{subset}, intended to subset the cancer record dataset. If \code{fun} does not
have an argument named \code{subset}, the dataset \code{x} is itself subset before
passing to \code{fun}. This is done separately for each entity number.

\code{loop_over_entity_columns} modifies or creates \code{arg_list[["by"]]} before
calling \code{fun}. \code{x} is not subset by this function. Therefore
\code{loop_over_entity_columns} is a more efficient but also less flexible
version of \code{loop_over_entity_numbers}.
}
