% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload.R
\name{nordcan_statistics_payload}
\alias{nordcan_statistics_payload}
\title{NORDCAN Statistics Payload}
\usage{
nordcan_statistics_payload(
  datasets = list(cancer_death_count_dataset = cancer_death_count_dataset,
    cancer_record_dataset = cancer_record_dataset, cancer_case_dataset =
    cancer_case_dataset)
)
}
\arguments{
\item{datasets}{\verb{[list]} (mandatory, no default)

a list containing datasets. the following are mandatory:
\itemize{
\item \code{cancer_record_dataset}: each row is a cancer record as output by
\link[nordcanpreprocessing:nordcan_incidence_dataset]{nordcanpreprocessing::nordcan_incidence_dataset}
\item \code{population_mean_size_dataset}: see
\link[nordcanpreprocessing:nordcan_population_mean_size_dataset]{nordcanpreprocessing::nordcan_population_mean_size_dataset}
\item \code{population_mortality_dataset}: see
\link[nordcanpreprocessing:nordcan_population_mortality_dataset]{nordcanpreprocessing::nordcan_population_mortality_dataset}
}

the following are optional and are computed within this function if not
supplied:
\itemize{
\item \code{cancer_death_count_dataset}: contains the numbers of cancer deaths; see
\link[nordcanpreprocessing:nordcan_cancer_death_count_dataset]{nordcanpreprocessing::nordcan_cancer_death_count_dataset}
}}
}
\description{
Compute all necessary statistics for NORDCAN into a single payload.
}
\examples{

nc_stats_payload <- nordcanepistats::nordcan_statistics_payload(
  datasets = list(
    cancer_record_dataset = my_crd,
    cancer_death_count_dataset = my_cdcd,
    population_mean_size_dataset = my_mps,
    population_mortality_dataset = my_md
  )
)
}
