% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload.R
\name{nordcan_statistics_tables_input_names}
\alias{nordcan_statistics_tables_input_names}
\title{NORDCAN Statistics Payload}
\usage{
nordcan_statistics_tables_input_names()
}
\arguments{
\item{datasets}{\verb{[list]} (mandatory, no default)

a list containing the followin datasets:
\itemize{
\item \code{cancer_record_dataset}: as output by
\link[nordcanpreprocessing:nordcan_processed_cancer_record_dataset]{nordcanpreprocessing::nordcan_processed_cancer_record_dataset}
\item \code{general_population_size_dataset}: see
\link[nordcanpreprocessing:nordcan_general_population_size_dataset]{nordcanpreprocessing::nordcan_general_population_size_dataset}
\item \code{national_population_life_table}: see
\link[nordcanpreprocessing:nordcan_national_population_life_table]{nordcanpreprocessing::nordcan_national_population_life_table}
\item \code{cancer_death_count_dataset}: dataset containing numbers of cancer deaths;
the creation of this dataset is explained in
\link[nordcanpreprocessing:nordcan_processed_cancer_death_count_dataset]{nordcanpreprocessing::nordcan_processed_cancer_death_count_dataset}
}}
}
\description{
Compute all necessary statistics for NORDCAN into a single payload.
}
\examples{

nc_stats <- nordcan_statistics_tables(
  datasets = list(
    cancer_record_dataset = my_crd,
    cancer_death_count_dataset = my_cdcd,
    general_population_size_dataset = my_gpsd,
    national_population_life_table = my_lt
  )
)
}
