% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_quality.R
\name{nordcanstat_imp_quality}
\alias{nordcanstat_imp_quality}
\title{Incidence, Mortality and Prevalence Quality Statistics}
\usage{
nordcanstat_imp_quality(
  x,
  cancer_death_count_dataset,
  by = NULL,
  subset = NULL,
  type = c("general", "exclusion")[1],
  subset_style = "zeros"
)
}
\arguments{
\item{x}{see \link[basicepistats:counts]{basicepistats::stat_count}}

\item{cancer_death_count_dataset}{\verb{[data.table]} (mandatory, no default)}

\item{by}{see \link[basicepistats:counts]{basicepistats::stat_count}}

\item{subset}{see \link[basicepistats:counts]{basicepistats::stat_count}}

\item{subset_style}{see \link[basicepistats:counts]{basicepistats::stat_count}}
}
\description{
Compute basic IMP quality statistics.
}
\seealso{
Other nordcanstat: 
\code{\link{nordcanstat_count}()},
\code{\link{nordcanstat_survival_quality}()},
\code{\link{nordcanstat_year_based_prevalent_patient_count}()}
}
\concept{nordcanstat}
