% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{nordcanstat_prevalent_subject_count}
\alias{nordcanstat_prevalent_subject_count}
\title{Prevalence}
\usage{
nordcanstat_prevalent_subject_count(
  x,
  by = NULL,
  subset = NULL,
  subset_style = "zeros"
)
}
\arguments{
\item{x}{see [basicepistats::stat_prevalent_subject_count]}

\item{by}{see [basicepistats::stat_prevalent_subject_count]}

\item{subset}{see [basicepistats::stat_prevalent_subject_count]}

\item{subset_style}{see [basicepistats::stat_prevalent_subject_count]}
}
\description{
Compute numbers of people alive at specific time points who have had
a cancer diagnosis in their past.
}
