% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_quality.R
\name{nordcanstat_survival_quality}
\alias{nordcanstat_survival_quality}
\title{Survival Quality Statsitics}
\usage{
nordcanstat_survival_quality(
  x,
  stratum_col_nms = NULL,
  subset = NULL,
  subset_style = "zeros"
)
}
\arguments{
\item{x}{see \link[basicepistats:counts]{basicepistats::stat_count}}

\item{subset}{see \link[basicepistats:counts]{basicepistats::stat_count}}

\item{subset_style}{see \link[basicepistats:counts]{basicepistats::stat_count}}

\item{by}{see \link[basicepistats:counts]{basicepistats::stat_count}}
}
\description{
Compute basic survival quality statistics.
}
\seealso{
Other nordcanstat: 
\code{\link{nordcanstat_count}()},
\code{\link{nordcanstat_year_based_prevalent_subject_count}()}
}
\concept{nordcanstat}
