% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{nordcanstat_year_based_prevalent_subject_count}
\alias{nordcanstat_year_based_prevalent_subject_count}
\title{Prevalence}
\usage{
nordcanstat_year_based_prevalent_subject_count(
  x,
  by = NULL,
  subset = NULL,
  subset_style = "zeros"
)
}
\arguments{
\item{x}{see \link[basicepistats:year_based_prevalence]{basicepistats::stat_year_based_prevalent_subject_count}}

\item{by}{see \link[basicepistats:year_based_prevalence]{basicepistats::stat_year_based_prevalent_subject_count}}

\item{subset}{see \link[basicepistats:year_based_prevalence]{basicepistats::stat_year_based_prevalent_subject_count}}

\item{subset_style}{see \link[basicepistats:year_based_prevalence]{basicepistats::stat_year_based_prevalent_subject_count}}
}
\description{
Compute numbers of people alive at specific time points who have had
a cancer diagnosis in their past.
}
\seealso{
Other nordcanstat: 
\code{\link{nordcanstat_count}()},
\code{\link{nordcanstat_survival_quality}()}
}
\concept{nordcanstat}
