% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nordcan_statistics_tables.R
\name{write_nordcan_statistics_tables}
\alias{write_nordcan_statistics_tables}
\alias{write_nordcan_statistics_tables_for_archive}
\alias{write_nordcan_statistics_tables_for_sending}
\title{Write NORDCAN Statistics Tables to Zip}
\usage{
write_nordcan_statistics_tables_for_archive(input)

write_nordcan_statistics_tables_for_sending(input)
}
\arguments{
\item{input}{a list of output of nordcan statistics tables}
}
\value{
a zip file named 'nordcan_statistics_tables.zip' will be save in
current work directory (which should be setup before using
\link[nordcancore:global_nordcan_settings]{nordcancore::set_global_nordcan_settings})
}
\description{
Write the output of \link{nordcan_statistics_tables} into a .zip file.
}
\details{
\itemize{
\item \code{write_nordcan_statistics_tables_for_archive} writes data.tables as .csv
character vectors as .txt files
}

\itemize{
\item \code{write_nordcan_statistics_tables_for_archive} writes data.tables as .csv
character vectors as .txt files
}
}
\examples{

\dontrun{
nordcancore::set_global_nordcan_settings(
  work_dir = ".",
  participant_name = "Norway",
  stat_cancer_record_count_first_year = 1954L,
  stat_prevalent_subject_count_first_year = 1954L,
  stat_cancer_death_count_first_year = 1954L,
  stat_survival_follow_up_first_year = 1954L
)


input <- list(log1 = letters, log2 = LETTERS[1:5],
               output1 = as.data.table(cars), output2 = as.data.table(CO2))


write_nordcan_statistics_tables_for_archive(input = input)
}
\dontrun{
nordcancore::set_global_nordcan_settings(
  work_dir = ".",
  participant_name = "Norway",
  stat_cancer_record_count_first_year = 1954L,
  stat_prevalent_subject_count_first_year = 1954L,
  stat_cancer_death_count_first_year = 1954L,
  stat_survival_follow_up_first_year = 1954L
)


input <- list(log1 = letters, log2 = LETTERS[1:5],
              output1 = as.data.table(cars), output2 = as.data.table(CO2))


write_nordcan_statistics_tables_for_archive(input = input)
}
}
