% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_by_entity}
\alias{compute_by_entity}
\alias{compute_by_entity_column}
\alias{nordcanstat_by_entity_column}
\title{Stratification by Entity}
\usage{
compute_by_entity(x, entities, fun, arg_list)

compute_by_entity_column(x, entities, fun, arg_list)

nordcanstat_by_entity_column(
  x,
  by,
  entities,
  subset,
  subset_style,
  basicepistats_fun
)
}
\arguments{
\item{x}{\verb{[data.table]} (mandatory, no default)

the NORDCAN cancer record dataset after preprocessing; see also the
underlying implementation \link[basicepistats:counts]{basicepistats::stat_count}}

\item{entities}{\verb{[NULL, integer]} (optional, default \code{NULL})

see \code{by} for how to stratify output by entity number.
\itemize{
\item \code{NULL}: output of
\code{nordcancore::nordcan_metadata_entity_no_set("all")}
is used
\item \code{integer}: these entity numbers are used
}}

\item{fun}{\verb{[function]} (mandatory, no default)

function that is applied to each entity / entity column; see Details}

\item{arg_list}{\verb{[list]} (mandatory, no default)

a \code{list} containing arguments to pass to \code{fun}; see Details}

\item{by}{see \link[basicepistats:counts]{basicepistats::stat_count}

exceptions to \verb{[basicepistats::stat_count]}:
\itemize{
\item if \code{by} is a character vector of column names in \code{x}, the allowed
combinations of those columns are retrieved using
\verb{[nordcancore::nordcan_metadata_column_level_space_dt]}, and the output
will include all of these combinations
\item additionally if \code{by} is a character vector and one element of it is
\code{"entity"}, then output is also stratified by entity number
(the entity numbers will be in column \code{entity} in output); the
entities to compute results by are supplied via \code{entities}
}}

\item{subset}{see the same argument in e.g. \link[basicepistats:counts]{basicepistats::stat_count}}

\item{subset_style}{see the same argument in e.g. \link[basicepistats:counts]{basicepistats::stat_count}}

\item{basicepistats_fun}{\verb{[function]} (mandatory, no default)

function from \strong{basicepistats}, e.g. \verb{[basicepistats::stat_count]};
this function will be run to produce results by entity number}
}
\value{
Either a \code{list} or a \code{data.table}. Initially results are collected into a
\code{list} where each element is the set of results for an individual entity
or entity column. If every element of this list is a \code{data.table},
then every element is combined into one long \code{data.table} where column
\code{"entity"} will identify the entity. In the case of
\code{compute_by_entity} the column \code{"entity"} is created and in the case of
\code{compute_by_entity_column} each entity column is renamed to \code{"entity"}
in the entity column-specific results before combining.
}
\description{
Compute statistics by entity number.
}
\details{
\code{compute_by_entity} effectively loops over each entity number given in
\code{entities} and computes the results separately for each.

\code{compute_by_entity_column} loops over each column defining entities in
\code{x}. For each such column, \code{arg_list[["by"]]} is created or modified to have
(also) the column in question, limited to only those entities supplied in
\code{entities}.

Both of these two functions are intended to be applied internally
by the \code{nordcanstat_} family of functions and not by the user.

\code{fun} is called using \code{arg_list} after the appropriate entity- or
entity column-specific modification using \verb{[base::do.call]}.

\code{compute_by_entity} modifies an element in \code{arg_list} named
\code{x}, intended to contain the dataset cancer record. \code{x} is subset into
each entity number separately before calling \code{fun}, i.e.
\code{arg_list[["x"]] <- x[has_first_entity_no, ]} for the first entity and so on.

\code{compute_by_entity_column} modifies or creates \code{arg_list[["by"]]} before
calling \code{fun}. \code{x} is not subset by this function. Therefore
\code{compute_by_entity_column} is a more efficient but also less flexible
version of \code{compute_by_entity}.
}
