% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload.R
\name{nordcan_statistics_payload}
\alias{nordcan_statistics_payload}
\title{NORDCAN Statistics Payload}
\usage{
nordcan_statistics_payload(datasets)
}
\arguments{
\item{datasets}{`[list]` (mandatory, no default)

a list containing datasets. the following are mandatory:
- `cancer_record_dataset`: each row is a cancer record as output by
  [nordcanpreprocessing::nordcan_incidence_dataset]
- `population_mean_size_dataset`: see
  [nordcanpreprocessing::nordcan_population_mean_size_dataset]
- `population_mortality_dataset`: see
  [nordcanpreprocessing::nordcan_population_mortality_dataset]

the following are optional and are computed within this function if not
supplied:
- `cancer_death_count_dataset`: contains the numbers of cancer deaths; see
  [nordcanpreprocessing::nordcan_cancer_death_count_dataset]}
}
\description{
Compute all necessary statistics for NORDCAN into a single payload.
}
\examples{

nc_stats_payload <- nordcanepistats::nordcan_statistics_payload(
  datasets = list(
    cancer_record_dataset = my_crd,
    cancer_death_count_dataset = my_cdcd,
    population_mean_size_dataset = my_mps,
    population_mortality_dataset = my_md
  )
)
}
