% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload.R
\name{nordcan_statistics_tables}
\alias{nordcan_statistics_tables}
\title{NORDCAN Statistics Payload}
\usage{
nordcan_statistics_tables(
  cancer_record_dataset,
  cancer_death_count_dataset,
  general_population_size_dataset,
  national_population_life_table,
  stata_exe_path
)
}
\arguments{
\item{cancer_record_dataset}{\verb{[data.table]} (mandatory, no default)

as output by
\link[nordcanpreprocessing:nordcan_processed_cancer_record_dataset]{nordcanpreprocessing::nordcan_processed_cancer_record_dataset}}

\item{cancer_death_count_dataset}{\verb{[data.table]} (mandatory, no default)
dataset containing numbers of cancer deaths; see Details}

\item{general_population_size_dataset}{\verb{[data.table]} (mandatory, no default)
the dataset of population sizes as per the call for data}

\item{national_population_life_table}{\verb{[data.table]} (mandatory, no default)
the life table as per the call for data}

\item{stata_exe_path}{\verb{[character]} (mandatory, no default)

pased to \link[nordcansurvival:nordcanstat_survival]{nordcansurvival::nordcanstat_survival}}
}
\description{
Compute all necessary statistics for NORDCAN into a single payload.
}
\details{
You need to form \code{cancer_death_count_dataset} yourself using the raw data
you have using one of two methods.
If you have a dataset of cancer death counts as described in the call for
data, do\preformatted{cdcd <- nordcanpreprocessing::nordcan_processed_cancer_death_count_dataset(
  my_raw_cdcd
)
}

where \code{my_raw_cdcd} is your dataset of cancer death counts as per the call
for data.

If you want to compute the counts using your cancer record dataset, do\preformatted{cdcd <- nordcanepistats::nordcanstat_count(
  processed_cancer_record_dataset,
  by = c("sex", "entity", "yoi", "region", "agegroup"),
  subset = died_from_cancer == TRUE
)
data.table::setnames(cdcd,
                     c("N", "yoi"), c("cancer_death_count", "year"))
}

where \code{processed_cancer_record_dataset} is your cancer record dataset after
processing
(see \link[nordcanpreprocessing:nordcan_processed_cancer_record_dataset]{nordcanpreprocessing::nordcan_processed_cancer_record_dataset}),
and in this example the information on who died of which cancer is identified
in the logical vector \code{died_from_cancer}, which you need to define. It should
be of length \code{nrow(processed_cancer_record_dataset)}. One person can
naturally only die once, so there can be at most one \code{TRUE} value per person.
}
\examples{

\dontrun{
nc_stats <- nordcan_statistics_tables(
 cancer_record_dataset = crd,
 cancer_death_count_dataset = cdcd,
 general_population_size_dataset = gpsd,
 national_population_life_table = lt,
 stata_exe_path = "stata.exe"
)
}
}
