% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{nordcanstat_count}
\alias{nordcanstat_count}
\title{Cancer Record Counts}
\usage{
nordcanstat_count(x, by = NULL, subset = NULL, subset_style = "zeros")
}
\arguments{
\item{x}{see \link[basicepistats:counts]{basicepistats::stat_count}}

\item{by}{see \link[basicepistats:counts]{basicepistats::stat_count}}

\item{subset}{see \link[basicepistats:counts]{basicepistats::stat_count}}

\item{subset_style}{see \link[basicepistats:counts]{basicepistats::stat_count}}
}
\description{
Compute counts of cancer records.
}
\seealso{
Other nordcanstat: 
\code{\link{nordcanstat_survival_quality}()},
\code{\link{nordcanstat_year_based_prevalent_subject_count}()}
}
\concept{nordcanstat}
